move parameterwithsetpoints to its own module
